+++ dart/lib/src/protocol/sentry_stack_trace.dart
@@ -1,0 +1,33 @@
+import 'sentry_stack_frame.dart';
+
+class SentryStackTrace {
+  List<SentryStackFrame> _frames;
+
+  List<SentryStackFrame> get frames => List.unmodifiable(_frames);
+
+  set frames(List<SentryStackFrame> frames) {
+    _frames = frames;
+  }
+
+  Map<String, String> _registers;
+
+  Map<String, String> get registers => Map.unmodifiable(_registers);
+
+  set registers(Map<String, String> registers) {
+    _registers = registers;
+  }
+
+  Map<String, dynamic> toJson() {
+    final json = <String, dynamic>{};
+
+    if (_frames?.isNotEmpty ?? false) {
+      json['frames'] = _frames.map((frame) => frame.toJson()).toList();
+    }
+
+    if (_registers?.isNotEmpty ?? false) {
+      json['registers'] = _registers;
+    }
+
+    return json;
+  }
+}